requirements.txt
mcp[cli]
openai
google-genai
python-dotenv
anthropic
loguru # or use logging
